requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pydantic
google-generativeai
playwright
beautifulsoup4
